Moved Page3Service to new folder, added spec
