+++ src/providers/page3-service/page3-service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@angular/core';
 import { BehaviorSubject } from 'rxjs';
 import 'rxjs/add/operator/map';
 
-export interface Page3Item {
+export class Page3Item {
   title: string;
   note: string;
   icon: string;
@@ -22,12 +22,12 @@ export class Page3Service {
 
   private createDataForService() {
     this.page3data = [];
-    for (let i = 1; i < 26; i++) {
-      this.page3data.push({
-        title: 'Item ' + i,
-        note: 'This is item #' + i,
-        icon: iconNames[Math.floor(Math.random() * iconNames.length)]
-      });
+    for (let i = 1; i <= 25; i++) {
+      let item = new Page3Item();
+      item.title = 'Item ' + i;
+      item.note = 'This is item #' + i;
+      item.icon = iconNames[Math.floor(Math.random() * iconNames.length)];
+      this.page3data.push(item);
     }
     this.page3Subject.next(this.page3data);
